Created methods in DataSource to return source lists; Added isRented property in Car class; Added logic in list available/rented car methods; Added logic in rent a car method; Created calculatePrice method; Added RentedCarHistory class;

+++ src/main/java/ro/jademy/carrental/model/cars/Dacia.java
@@ -4,7 +4,7 @@ import ro.jademy.carrental.model.Car;
 
 public class Dacia extends Car {
 
-            public Dacia(String make, String model, String fuelType, String carType, String transmissionType, String carId, int nrOfSeats, int nrOfDoors, double pricePerDay) {
+            public Dacia(String make, String model, String fuelType, String carType, String transmissionType, String carId, int nrOfSeats, int nrOfDoors, int pricePerDay) {
             super(make, model, fuelType, carType, transmissionType, carId, nrOfSeats, nrOfDoors, pricePerDay);
         }
