+++ src/main/java/ro/jademy/carrental/model/cars/Fiat.java
@@ -4,7 +4,7 @@ import ro.jademy.carrental.model.Car;
 
 public class Fiat extends Car {
 
-    public Fiat(String make, String model, String fuelType, String carType, String transmissionType, String carId, int nrOfSeats, int nrOfDoors, double pricePerDay) {
+    public Fiat(String make, String model, String fuelType, String carType, String transmissionType, String carId, int nrOfSeats, int nrOfDoors, int pricePerDay) {
         super(make, model, fuelType, carType, transmissionType, carId, nrOfSeats, nrOfDoors, pricePerDay);
     }
 }
