+++ src/main/java/ro/jademy/carrental/model/cars/Nissan.java
@@ -3,7 +3,7 @@ package ro.jademy.carrental.model.cars;
 import ro.jademy.carrental.model.Car;
 
 public class Nissan extends Car {
-    public Nissan(String make, String model, String fuelType, String carType, String transmissionType, String carId, int nrOfSeats, int nrOfDoors, double pricePerDay) {
+    public Nissan(String make, String model, String fuelType, String carType, String transmissionType, String carId, int nrOfSeats, int nrOfDoors, int pricePerDay) {
         super(make, model, fuelType, carType, transmissionType, carId, nrOfSeats, nrOfDoors, pricePerDay);
     }
 }
